Add files via upload
Further improvements in FT4 & FT8 signal bandwidths with much more precise gaussian transitions. The slow exp function calcs using double type variables have been replaced with an interpolated look up table and special 64 bit integer calculations. This fully capitalises on the capabilities of the AD9851 DDS chip to make rapid phase continuous frequency changes with 0.0419Hz resolution that can now be made up to every 60 microsecs.

+++ TXlink4_Nano/AD9851.h
@@ -72,7 +72,7 @@ public:
   uint32_t frequencyFTW(uint32_t freq_x10) const
   {
     // 32-bit path: (freq_x10 * (reciprocal / 10) + rounding_term) >> shift
-    return ((uint64_t)freq_x10 * (reciprocal / 10) + (reciprocal/16/10)) >> shift;
+    return (uint32_t) (((uint64_t)freq_x10 * (reciprocal / 10) + (reciprocal/16/10)) >> shift);
   }
 
   void setFrequencyHz(uint32_t freq, bool PowerUp)
